requirements.txt
akracer==0.0.13
akshare==1.13.53
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.8.1
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.1.1
cryptography==44.0.3
customtkinter==5.2.2
cycler==0.12.1
Cython==3.1.0
darkdetect==0.8.0
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
docutils==0.20.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Faker==35.2.2
favicon==0.7.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
folium==0.18.0
fonttools==4.51.0
frozendict==2.4.2
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
glcontext==2.5.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
htbuilder==0.9.0
html5lib==1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.0
jsonpath==0.82.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
moderngl==5.10.0
mplfinance==0.12.10b0
multitasking==0.0.11
narwhals==1.21.1
numpy==1.24.4
oauth2client==4.1.3
objloader==0.2.0
openai==1.77.0
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pandas-datareader==0.10.0
peewee==3.17.3
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==6.0.1
prometheus_client==0.21.1
protobuf==5.29.2
py-mini-racer==0.6.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
pymdown-extensions==10.15
PyOpenGL==3.1.7
pyparsing==3.1.2
pypinyin==0.51.0
Pyrebase4==4.8.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.15.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.40.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-geolocation==0.0.10
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_carousel==1.1.2
streamlit_folium==0.22.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
validators==0.34.0
virtualenv==20.26.3
webencodings==0.5.1
Werkzeug==3.0.6
xlrd==2.0.1
xyzservices==2025.4.0
yfinance==0.2.38
zipp==3.20.2
